CDROM: ResponseBuffer EnqueueRange from array params to stackalloc spans

+++ ProjectPSX/Devices/CDROM/TrackBuilder.cs
@@ -11,12 +11,12 @@ namespace ProjectPSX.Devices.CdRom {
 
             public String file { get; private set; }
             public long size { get; private set; }
-            public int number { get; private set; }
+            public byte number { get; private set; }
             public int lba { get; private set; }
             public int lbaStart { get; private set; }
             public int lbaEnd { get; private set; }
 
-            public Track(String file, long size, int number, int lba, int lbaStart, int lbaEnd) {
+            public Track(String file, long size, byte number, int lba, int lbaStart, int lbaEnd) {
                 this.file = file;
                 this.size = size;
                 this.number = number;
@@ -32,7 +32,7 @@ namespace ProjectPSX.Devices.CdRom {
             String dir = Path.GetDirectoryName(cue);
             String line;
             int lbaCounter = 0;
-            int number = 0;
+            byte number = 0;
             using StreamReader cueFile = new StreamReader(cue);
             while ((line = cueFile.ReadLine()) != null) {
                 if (line.StartsWith("FILE")) {
@@ -70,7 +70,7 @@ namespace ProjectPSX.Devices.CdRom {
             int lba = (int)(size / BytesPerSectorRaw);
             int lbaStart = 150; // 150 frames (2 seconds) offset from track 1
             int lbaEnd = lba;
-            int number = 1;
+            byte number = 1;
 
             tracks.Add(new Track(file, size, number, lba, lbaStart, lbaEnd));
